requirements.txt
torch
opencv-python
numpy
joblib
Pillow
transformers
ultralytics
scikit-learn
protobuf<4
typing_extensions>=4.5.0
hypothesis
pytest
pytest-cov
pyyaml
matplotlib
tqdm
flask
flask-cors
flask-socketio
python-socketio
